build mixin for automatically changing pixels to percentages
mixin is called pixToPer and specifically replaces a width: 100px
for example, with a formula of width: 100 / whatever the container
width is * 100% which returns a percentage width based on the
formula. Having problems in fluid design with the get started
button. It's getting properties for pixel width from somewhere
other than the css that I'm finding.