Major adjustments to the code
Finished adjusting the image in the middle. Now is resizing according to the viewport (desktop, tablet, mobile.

Finished setting up the logo at the navigation menu for the tablet layout. Worked on my HTML file to be able to put the logo on top of the page. The image tab was inside the navigation tab.

The only thing missing is changing the text to rem units.